Revert "Make frameworks/av 64-bit compatible"
This reverts commit ee8d3a3778d5dcafaf07cb2254096e302157b3a9.

+++ media/libstagefright/codecs/m4v_h263/enc/src/motion_comp.cpp
@@ -1959,7 +1959,7 @@ void PutSkippedBlock(UChar *rec, UChar *prev, Int lx)
         dst += offset;
         src += offset;
     }
-    while ((uintptr_t)src < (uintptr_t)end);
+    while ((UInt)src < (UInt)end);
 
     return ;
 }
